Make infoBackupDataAnnotationSet() more efficient.
Make a copy of the manifest annotations so they don't get modified in the manifest.

Update logic to avoid leaking a KV when one already exists.

Update variable names to more accurately reflect what is happening. Reorder if to be clearer.

Remove useless check for kvKeyExists() which is already done in kvRemove().