Fix incorrect storage for rendering errors.
Coverity complained of a copy/paste error here, but the actual error was what it took to be the correct "copied from" code.

In any case, the prior code would have blown up as soon as a new error type was added. Fix by updating to the writable build storage.