Shim sckClientOpen() instead of ioSessionFd().
The problem with shimming ioSessionFd() is that a socket connection will still be made to the SSH server. This is not a good idea for unit tests and make the test require a container to run, which limits where we can test.

Instead shim sckClientOpen() which bypasses the socket connection but still sets the arbitrary fd.