Write block totals instead of a block list in super blocks.
Within a super block the blocks must be consecutive so there is no need to write them out explicitly.

However, the first block might be offset after a continuation or if the first blocks of a super block were overridden by a later super block. In either of these cases the offset is stored after the block total.

The space savings are 11-14% for checksum sizes 6-7.